[refactor] #6 information -> notice로 파일명 리팩토링
