+++ app/src/main/AndroidManifest.xml
@@ -25,7 +25,7 @@
         </activity>
 
         <activity
-            android:name=".presentation.employee.home.information.HomeInformationActivity"
+            android:name=".presentation.employee.home.notice.HomeNoticeActivity"
             android:exported="true"
             android:screenOrientation="portrait"
             tools:ignore="LockedOrientationActivity" >
