--- app/src/main/java/com/omg/cafebara/presentation/employee/home/information/HomeInformationViewModel.kt
@@ -1,6 +1,0 @@
-package com.omg.cafebara.presentation.employee.home.information
-
-import androidx.lifecycle.ViewModel
-
-class HomeInformationViewModel: ViewModel() {
-}
