+++ app/src/main/java/com/omg/cafebara/presentation/employee/home/notice/HomeNoticeViewModel.kt
@@ -1,0 +1,6 @@
+package com.omg.cafebara.presentation.employee.home.notice
+
+import androidx.lifecycle.ViewModel
+
+class HomeNoticeViewModel: ViewModel() {
+}
