+++ app/src/main/res/layout/activity_home_notice.xml
@@ -5,7 +5,7 @@
     <data>
         <variable
             name="viewModel"
-            type="com.omg.cafebara.presentation.employee.home.information.HomeInformationViewModel" />
+            type="com.omg.cafebara.presentation.employee.home.notice.HomeNoticeViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
