+++ app/src/main/res/layout/fragment_home_notice_detail.xml
