+++ app/src/main/res/layout/item_home_notice.xml
